fixed bugs, implemented database, improved style
added the database functionality in the shopping cart, each user will now have an `amounts` key, and in that key will be a key-value pair for each item amount. That information will be updated upon clicking the add/remove buttons and on logging in.

Added images and improved styling of the UI, Removed the unnecessary description popup, and passed the user id as an intent.